pyproject.toml
[project]
name = "ai-agent-challenge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.76",
    "langchain-groq>=0.3.8",
    "langchain-huggingface>=0.3.1",
    "langgraph>=0.6.7",
    "numpy>=2.3.3",
    "pandas>=2.3.2",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
]
